% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{break_names}
\alias{break_names}
\title{break_names}
\usage{
break_names(the_names, long_name_limit = 46)
}
\arguments{
\item{the_names}{A vector of names}

\item{long_name_limit}{The limit to define a long name (defautl 46 chars.)}
}
\value{
The list of adapted names
}
\description{
Introduce a line break in the middle of a long name.
}
