% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_reactome_url}
\alias{check_reactome_url}
\title{check_reactome_url}
\usage{
check_reactome_url(reactome_url)
}
\arguments{
\item{reactome_url}{character The URL to test. If \code{NULL} the URL
is retrieved from the options.}
}
\value{
character The potentially cleaned / retrieved URL with a trailing "/"
}
\description{
Makes sure the passed URL is valid. If not URL is passed, the one
stored in the options is retrieved
}
