% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_request.R
\name{checkRequestValidity}
\alias{checkRequestValidity}
\title{Check's if a ReactomeAnalysisRequest object is valid}
\usage{
checkRequestValidity(object)
}
\arguments{
\item{object}{The request object to check.}
}
\value{
TRUE if the object is valid or a string with the reason why it is not
}
\description{
Check's if a ReactomeAnalysisRequest object is valid
}
