% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result.R
\name{convert_reactome_result}
\alias{convert_reactome_result}
\title{Convert the REACTOME JSON result to a ReactomeAnalysisResult object}
\usage{
convert_reactome_result(reactome_result)
}
\arguments{
\item{reactome_result}{The JSON result already converted to R objects (name list)}
}
\value{
A \code{\link{ReactomeAnalysisResult}} object
}
\description{
Convert the REACTOME JSON result to a ReactomeAnalysisResult object
}
