% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_frame_as_string}
\alias{data_frame_as_string}
\title{Converts a data.frame to a string representation}
\usage{
data_frame_as_string(data)
}
\arguments{
\item{data}{The data.frame to convert}
}
\value{
A string representing the passed data.frame
}
\description{
A data.frame is converted into a single string using
`\\t` (the characters, not tab) as field delimiter and
`\\n` (the characters, not newline) as line delimiter
}
