% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_data.R
\name{fetch_public_data}
\alias{fetch_public_data}
\title{fetch_public_data}
\usage{
fetch_public_data(dataset_entry, reactome_url)
}
\arguments{
\item{dataset_entry}{The entry of the respective dataset as returned by 
the \code{\link{find_public_datasets}} function.}

\item{reactome_url}{URL of the Reactome API Server. Overwrites the URL set in the 'reactome_gsa.url' option.
Specific ports can be set using the standard URL specification (for example http://your.service:1234)}
}
\value{
The loaded data as an ExpressionSet object.
}
\description{
Loads an already available public dataset from ReactomeGSA
and returns it as a Biobase::ExpressionSet object.
}
