% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata.R
\name{generate_metadata,data.frame-method}
\alias{generate_metadata,data.frame-method}
\title{Generate metadata}
\usage{
\S4method{generate_metadata}{data.frame}(pseudo_bulk_data)
}
\arguments{
\item{pseudo_bulk_data}{Pseudobulk data generated from the
\code{\link{generate_pseudo_bulk_data}} function}
}
\value{
Returns metadata table for later use
}
\description{
Generate metadata
}
