% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata.R
\name{generate_metadata}
\alias{generate_metadata}
\title{generate_metadata}
\usage{
generate_metadata(pseudo_bulk_data)
}
\arguments{
\item{pseudo_bulk_data}{Pseudobulk data generated from the 
\code{\link{generate_pseudo_bulk_data}} function}
}
\value{
Metadata table for later use
}
\description{
The pseudobulk data is generated using the 
 \code{\link{generate_pseudo_bulk_data}} function.
}
\examples{
# Example pseudobulk data
pseudo_bulk_data <- data.frame(
  sample1_groupA = c(10, 20, 30),
  sample2_groupA = c(15, 25, 35),
  sample3_groupB = c(5, 10, 15)
)

# Generate metadata from pseudobulk data
metadata <- generate_metadata(pseudo_bulk_data)

}
\seealso{
\code{\link{generate_pseudo_bulk_data}} for
         generating pseudobulk data.
}
