% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_bulk_data.R
\name{generate_pseudo_bulk_data,Seurat-method}
\alias{generate_pseudo_bulk_data,Seurat-method}
\title{generate_pseudo_bulk_data - Seurat}
\usage{
\S4method{generate_pseudo_bulk_data}{Seurat}(
  object,
  group_by = NULL,
  split_by = "random",
  k_variable = 4
)
}
\arguments{
\item{object}{The object to analyse.}

\item{group_by}{entry in metadata table, based on these cluster annotation pseudo bulk is performed}

\item{split_by}{variable -> split by a variable within the metadata; k must be a string
random -> splits based on a random number; k must be a number
Louvain, Louvain_multilevel, SLM, Leiden -> subclusters k must be a list with [resolution, cluster_1, cluster_2]}

\item{k_variable}{variable dependent on the split_by}
}
\value{
returns pseudo bulk generated data
}
\description{
Generate Pseudo Bulk Data for Seurat Objects
}
