% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_data.R
\name{get_dataset_loading_status}
\alias{get_dataset_loading_status}
\title{Retrieves the status of the submitted dataset loading request}
\usage{
get_dataset_loading_status(loading_id, reactome_url = NULL)
}
\arguments{
\item{loading_id}{The dataset loading process' id}

\item{reactome_url}{URL of the Reactome API Server. Overwrites the URL set in the 'reactome_gsa.url' option.
Specific ports can be set using the standard URL specification (for example http://your.service:1234)}
}
\value{
A list containing the \code{id}, \code{status} (can be "running", "complete", "failed"),
        \code{description}, and \code{completed} (numeric between 0 - 1)
}
\description{
Retrieves the status of the submitted dataset loading request
}
