% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_fc_for_dataset}
\alias{get_fc_for_dataset}
\title{get_fc_for_dataset}
\usage{
get_fc_for_dataset(dataset, pathway_result)
}
\arguments{
\item{dataset}{Name of the dataset to retrieve the fold changes for.}

\item{pathway_result}{The data.frame created by the \code{pathways} function.}
}
\value{
A vector of fold-changes
}
\description{
Retrieve the fold-changes for all pathways of the defined dataset
}
