% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_is_sig_dataset}
\alias{get_is_sig_dataset}
\title{get_is_sig_dataset}
\usage{
get_is_sig_dataset(dataset, pathway_result)
}
\arguments{
\item{dataset}{Name of the dataset}

\item{pathway_result}{data.frame created by the \code{pathways} function}
}
\value{
A vector with 3=non-significant, 2=p<=0.05, 1=p<0.01
}
\description{
Determines how significant a pathway is across the datasets. Returns the
lowest significance.
}
