% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_reactome_analysis_result}
\alias{get_reactome_analysis_result}
\title{Retrieves the result of the submitted analysis using \code{\link{perform_reactome_analysis}}}
\usage{
get_reactome_analysis_result(analysis_id, reactome_url = NULL)
}
\arguments{
\item{analysis_id}{The running analysis' id}

\item{reactome_url}{URL of the REACTOME API Server. Overwrites the URL set in the 'reactome_gsa.url' option.
Specific ports can be set using the standard URL specification (for example http://your.service:1234)}
}
\value{
The result object
}
\description{
The result is only available if \code{\link{get_reactome_analysis_status}} indicates that the
analysis is complete.
}
