% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva_plotting_functions.R
\name{is_gsva_result}
\alias{is_gsva_result}
\title{is_gsva_result}
\usage{
is_gsva_result(object)
}
\arguments{
\item{object}{A \code{\link{ReactomeAnalysisResult}} object}
}
\value{
Boolean indicating whether the object is a GSVA result.
}
\description{
is_gsva_result
}
