% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result.R
\name{names,ReactomeAnalysisResult-method}
\alias{names,ReactomeAnalysisResult-method}
\title{ReactomeAnalysisResult - names}
\usage{
\S4method{names}{ReactomeAnalysisResult}(x)
}
\arguments{
\item{x}{ReactomeAnalysisResult.}
}
\value{
character vector with the names of the contained datasets
}
\description{
Retrieves the names of the contained datasets within an \code{\link{ReactomeAnalysisResult}}
object.
}
\examples{
# load an example result object
library(ReactomeGSA.data)
data(griss_melanoma_result)

# get the names of the available datasets
names(griss_melanoma_result)
}
\seealso{
Other ReactomeAnalysisResult functions:
\code{\link[=get_result]{get_result()}},
\code{\link[=open_reactome]{open_reactome()}},
\code{\link[=pathways]{pathways()}},
\code{\link[=plot_correlations]{plot_correlations()}},
\code{\link[=plot_gsva_heatmap]{plot_gsva_heatmap()}},
\code{\link[=plot_gsva_pathway]{plot_gsva_pathway()}},
\code{\link[=plot_heatmap]{plot_heatmap()}},
\code{\link[=plot_volcano]{plot_volcano()}},
\code{\link[=reactome_links]{reactome_links()}},
\code{\link[=result_types]{result_types()}}
}
\concept{ReactomeAnalysisResult functions}
