% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result.R
\name{open_reactome}
\alias{open_reactome}
\title{open_reactome}
\usage{
open_reactome(x, ...)
}
\arguments{
\item{x}{ReactomeAnalysisResult.}

\item{...}{Additional parameters passed to downstream functions.}
}
\value{
The opened link
}
\description{
Opens the specified Reactome visualization in the system's default browser.
}
\examples{
# Note: This function only works with a newly created result
# since the visualization links only stay active for 7 days

# load an example result
library(ReactomeGSA.data)
data(griss_melanoma_result)

# get the reactome link - this does only work
# with new results
# open_reactome(griss_melanoma_result)
}
\seealso{
Other ReactomeAnalysisResult functions:
\code{\link[=get_result]{get_result()}},
\code{\link{names,ReactomeAnalysisResult-method}},
\code{\link[=pathways]{pathways()}},
\code{\link[=plot_correlations]{plot_correlations()}},
\code{\link[=plot_gsva_heatmap]{plot_gsva_heatmap()}},
\code{\link[=plot_gsva_pathway]{plot_gsva_pathway()}},
\code{\link[=plot_heatmap]{plot_heatmap()}},
\code{\link[=plot_volcano]{plot_volcano()}},
\code{\link[=reactome_links]{reactome_links()}},
\code{\link[=result_types]{result_types()}}
}
\concept{ReactomeAnalysisResult functions}
