% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_volcano,ReactomeAnalysisResult-method}
\alias{plot_volcano,ReactomeAnalysisResult-method}
\title{ReactomeAnalysisResult - plot_volcano}
\usage{
\S4method{plot_volcano}{ReactomeAnalysisResult}(x, dataset = 1, ...)
}
\arguments{
\item{x}{ReactomeAnalysisResult. The analysis result to plot the volcano plot for.}

\item{dataset}{The name or index of the dataset to plot (first one by default).}

\item{...}{Additional parameters for specific implementations.}
}
\value{
A ggplot2 plot object representing the volcano plot.
}
\description{
Creates a volcano plot for the pathway analysis result. Every point represents one
pathway, the x-axis the log fold-change and the y-axis the adjusted p-value (-log10).
}
\details{
This function is only available for GSA-based analysis results.
}
\examples{
# load an example result
library(ReactomeGSA.data)
data(griss_melanoma_result)

# create the volcano plot for the first dataset
plot_obj <- plot_volcano(griss_melanoma_result)

# display the plot using `print(plot_obj)`
}
\seealso{
Other ReactomeAnalysisResult functions:
\code{\link[=get_result]{get_result()}},
\code{\link{names,ReactomeAnalysisResult-method}},
\code{\link[=open_reactome]{open_reactome()}},
\code{\link[=pathways]{pathways()}},
\code{\link[=plot_correlations]{plot_correlations()}},
\code{\link[=plot_gsva_heatmap]{plot_gsva_heatmap()}},
\code{\link[=plot_gsva_pathway]{plot_gsva_pathway()}},
\code{\link[=plot_heatmap]{plot_heatmap()}},
\code{\link[=reactome_links]{reactome_links()}},
\code{\link[=result_types]{result_types()}}
}
