% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_request.R
\name{remove_dataset}
\alias{remove_dataset}
\title{remove_dataset}
\usage{
remove_dataset(x, dataset_name)
}
\arguments{
\item{x}{The \code{\link{ReactomeAnalysisRequest}} to remove the dataset from}

\item{dataset_name}{character The dataset's name}
}
\value{
The updated \code{\link{ReactomeAnalysisRequest}}
}
\description{
Remove the dataset from the \code{\link{ReactomeAnalysisRequest}} object.
}
