% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result.R
\docType{methods}
\name{show,ReactomeAnalysisResult-method}
\alias{show,ReactomeAnalysisResult-method}
\title{show - ReactomeAnalysisResult}
\usage{
\S4method{show}{ReactomeAnalysisResult}(object)
}
\arguments{
\item{object}{ReactomeAnalysisResult.}
}
\value{
character classname of the object
}
\description{
Displays basic information about the \code{\link{ReactomeAnalysisResult}}
object.
}
\examples{
library(ReactomeGSA.data)
data(griss_melanoma_result)

show(griss_melanoma_result)
}
