% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_bulk_data.R
\name{split_clustering}
\alias{split_clustering}
\title{method implementation subclustering}
\usage{
split_clustering(seurat_object, group_by, res, alg, cluster1, cluster2)
}
\arguments{
\item{seurat_object}{The Seurat object to analyse.}

\item{group_by}{entry in metadata table, based on these cluster annotation pseudo bulk is performed}

\item{res}{The clustering resolution to use.}

\item{alg}{Seurat subclustering algorithm id}

\item{cluster1}{cluster to subcluster}

\item{cluster2}{cluster to subcluster}
}
\value{
returns pseudo bulk generated data
}
\description{
method implementation subclustering
}
