% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_bulk_data.R
\name{split_random_sce}
\alias{split_random_sce}
\title{split SCE Object with random pooling}
\usage{
split_random_sce(sce_object, group_by, k_variable)
}
\arguments{
\item{sce_object}{The SingleCellExperiment object to analyse.}

\item{group_by}{entry in metadata table, based on these cluster annotation pseudo bulk is performed}

\item{k_variable}{number of pools that should be created}
}
\value{
returns pseudo bulk generated data
}
\description{
split SCE Object with random pooling
}
