% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_bulk_data.R
\name{split_variable}
\alias{split_variable}
\title{split Seurat object by variable}
\usage{
split_variable(seurat_object, group_by, k_variable)
}
\arguments{
\item{seurat_object}{The Seurat object to analyse.}

\item{group_by}{entry in metadata table, based on these cluster annotation pseudo bulk is performed}

\item{k_variable}{variable dependent on the split_by -> meta data entry}
}
\value{
returns pseudo bulk generated data
}
\description{
split Seurat object by variable
}
