% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_bulk_data.R
\name{split_variable_random}
\alias{split_variable_random}
\title{split Seurat object by random pooling}
\usage{
split_variable_random(seurat_object, group_by, k_variable)
}
\arguments{
\item{seurat_object}{The Seurat object to analyse.}

\item{group_by}{entry in metadata table, based on these cluster annotation pseudo bulk is performed}

\item{k_variable}{number of random pools}
}
\value{
returns pseudo bulk generated data
}
\description{
split Seurat object by random pooling
}
