% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_request.R
\name{start_reactome_analysis}
\alias{start_reactome_analysis}
\title{Start Reactome Analysis}
\usage{
start_reactome_analysis(request, reactome_url = NULL)
}
\arguments{
\item{request}{\code{\link{ReactomeAnalysisRequest}} object to submit.}

\item{reactome_url}{URL of the Reactome API Server. Overwrites the URL set in the 'reactome_gsa.url' option.
Specific ports can be set using the standard URL specification (for example http://your.service:1234)}
}
\value{
character The analysis job's id.
}
\description{
Submits a \code{\link{ReactomeAnalysisRequest}} to the Reactome Analysis Service API and
returns the analysis id of the submitted job.
}
