% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_data.R
\name{wait_for_loading_dataset}
\alias{wait_for_loading_dataset}
\title{wait_for_loading_dataset}
\usage{
wait_for_loading_dataset(request, verbose, reactome_url)
}
\arguments{
\item{request}{The httr request object of the dataset loading request.}

\item{verbose}{If set to \code{TRUE}, the progress is displayed as a status bar.}

\item{reactome_url}{URL of the Reactome API Server. Overwrites the URL set in the 'reactome_gsa.url' option.
Specific ports can be set using the standard URL specification (for example http://your.service:1234)}
}
\description{
This function loops until the dataset is available. If
verbose is set to \code{TRUE}, the progress is displayed
in a status bar.
}
