% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_reactome_methods}
\alias{get_reactome_methods}
\title{Returns all available analysis methods from the REACTOME service}
\usage{
get_reactome_methods(print_methods = TRUE, return_result = FALSE,
  reactome_url = NULL)
}
\arguments{
\item{print_methods}{If set to \code{TRUE} (default) a (relatively) nice formatted version of the result is printed.}

\item{return_result}{If set to \code{TRUE}, the result is returned as a data.frame (see below)}

\item{reactome_url}{URL of the REACTOME API Server. Overwrites the URL set in the 'reactome_gsa.url' option.
Specific ports can be set using the standard URL specification (for example http://your.service:1234)}
}
\value{
If \code{return_result} is set to \code{TRUE}, a data.frame with one row per method. Each method has a name, description, and
        (optional) a list of parameters. Parameters again have a name, type, and description.
}
\description{
Returns all available analysis methods from the REACTOME service
}
\examples{
# retrieve the available methods only in an object
available_methods <- get_reactome_methods(print_methods = FALSE, return_result = TRUE)

# print all method names
available_methods$name

# list all parameters for the first method
first_method_parameters <- available_methods[1, "parameters"]
first_method_parameters

# simply print the available methods
get_reactome_methods()
}
\seealso{
Other Reactome Service functions: \code{\link{get_reactome_data_types}}
}
\author{
Johannes Griss
}
\concept{Reactome Service functions}
